run mbnet int8
---
mq-quad h616 debian aarch64 

| optlevel | before(ms) | after(ms) |
| -- | -- | -- |
| 0 | 36.684 | 31.053 |
| 1 | 29.551 | 22.682 |

---
bananapi-m2-zero2 h2+ armbian armv7

| optlevel | before(ms) | after(ms) |
| -- | -- | -- |
| 0 | 64.991 | 58.798 |
| 1 | 55.251 | 48.067 |